REF: misplaced tests
