+++ pandas/tests/io/formats/test_format.py
@@ -1,26 +1,20 @@
 """
 Test output formatting for Series/DataFrame, including to_string & reprs
 """
-from contextlib import nullcontext
 from datetime import (
     datetime,
-    time,
     timedelta,
 )
 from io import StringIO
 import itertools
-import locale
-from operator import methodcaller
 from pathlib import Path
 import re
 from shutil import get_terminal_size
 import sys
 import textwrap
 
-import dateutil
 import numpy as np
 import pytest
-import pytz
 
 import pandas as pd
 from pandas import (
@@ -40,13 +34,6 @@ import pandas._testing as tm
 
 from pandas.io.formats import printing
 import pandas.io.formats.format as fmt
-
-
-def get_local_am_pm():
-    """Return the AM and PM strings returned by strftime in current locale."""
-    am_local = time(1).strftime("%p")
-    pm_local = time(13).strftime("%p")
-    return am_local, pm_local
 
 
 @pytest.fixture(params=["string", "pathlike", "buffer"])
@@ -3318,243 +3305,6 @@ class TestDatetime64Formatter:
         result = fmt._Datetime64TZFormatter(x).get_result()
         assert result[0].strip() == "2999-01-01 00:00:00-08:00"
         assert result[1].strip() == "2999-01-02 00:00:00-08:00"
-
-
-class TestNaTFormatting:
-    def test_repr(self):
-        assert repr(NaT) == "NaT"
-
-    def test_str(self):
-        assert str(NaT) == "NaT"
-
-
-class TestPeriodIndexFormat:
-    def test_period_format_and_strftime_default(self):
-        per = pd.PeriodIndex([datetime(2003, 1, 1, 12), None], freq="h")
-
-        # Default formatting
-        msg = "PeriodIndex.format is deprecated"
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            formatted = per.format()
-        assert formatted[0] == "2003-01-01 12:00"  # default: minutes not shown
-        assert formatted[1] == "NaT"
-        # format is equivalent to strftime(None)...
-        assert formatted[0] == per.strftime(None)[0]
-        assert per.strftime(None)[1] is np.nan  # ...except for NaTs
-
-        # Same test with nanoseconds freq
-        per = pd.period_range("2003-01-01 12:01:01.123456789", periods=2, freq="ns")
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            formatted = per.format()
-        assert (formatted == per.strftime(None)).all()
-        assert formatted[0] == "2003-01-01 12:01:01.123456789"
-        assert formatted[1] == "2003-01-01 12:01:01.123456790"
-
-    def test_period_custom(self):
-        # GH#46252 custom formatting directives %l (ms) and %u (us)
-        msg = "PeriodIndex.format is deprecated"
-
-        # 3 digits
-        per = pd.period_range("2003-01-01 12:01:01.123", periods=2, freq="ms")
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            formatted = per.format(date_format="%y %I:%M:%S (ms=%l us=%u ns=%n)")
-        assert formatted[0] == "03 12:01:01 (ms=123 us=123000 ns=123000000)"
-        assert formatted[1] == "03 12:01:01 (ms=124 us=124000 ns=124000000)"
-
-        # 6 digits
-        per = pd.period_range("2003-01-01 12:01:01.123456", periods=2, freq="us")
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            formatted = per.format(date_format="%y %I:%M:%S (ms=%l us=%u ns=%n)")
-        assert formatted[0] == "03 12:01:01 (ms=123 us=123456 ns=123456000)"
-        assert formatted[1] == "03 12:01:01 (ms=123 us=123457 ns=123457000)"
-
-        # 9 digits
-        per = pd.period_range("2003-01-01 12:01:01.123456789", periods=2, freq="ns")
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            formatted = per.format(date_format="%y %I:%M:%S (ms=%l us=%u ns=%n)")
-        assert formatted[0] == "03 12:01:01 (ms=123 us=123456 ns=123456789)"
-        assert formatted[1] == "03 12:01:01 (ms=123 us=123456 ns=123456790)"
-
-    def test_period_tz(self):
-        # Formatting periods created from a datetime with timezone.
-        msg = r"PeriodIndex\.format is deprecated"
-        # This timestamp is in 2013 in Europe/Paris but is 2012 in UTC
-        dt = pd.to_datetime(["2013-01-01 00:00:00+01:00"], utc=True)
-
-        # Converting to a period looses the timezone information
-        # Since tz is currently set as utc, we'll see 2012
-        with tm.assert_produces_warning(UserWarning, match="will drop timezone"):
-            per = dt.to_period(freq="h")
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            assert per.format()[0] == "2012-12-31 23:00"
-
-        # If tz is currently set as paris before conversion, we'll see 2013
-        dt = dt.tz_convert("Europe/Paris")
-        with tm.assert_produces_warning(UserWarning, match="will drop timezone"):
-            per = dt.to_period(freq="h")
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            assert per.format()[0] == "2013-01-01 00:00"
-
-    @pytest.mark.parametrize(
-        "locale_str",
-        [
-            pytest.param(None, id=str(locale.getlocale())),
-            "it_IT.utf8",
-            "it_IT",  # Note: encoding will be 'ISO8859-1'
-            "zh_CN.utf8",
-            "zh_CN",  # Note: encoding will be 'gb2312'
-        ],
-    )
-    def test_period_non_ascii_fmt(self, locale_str):
-        # GH#46468 non-ascii char in input format string leads to wrong output
-
-        # Skip if locale cannot be set
-        if locale_str is not None and not tm.can_set_locale(locale_str, locale.LC_ALL):
-            pytest.skip(f"Skipping as locale '{locale_str}' cannot be set on host.")
-
-        # Change locale temporarily for this test.
-        with tm.set_locale(locale_str, locale.LC_ALL) if locale_str else nullcontext():
-            # Scalar
-            per = pd.Period("2018-03-11 13:00", freq="h")
-            assert per.strftime("%y é") == "18 é"
-
-            # Index
-            per = pd.period_range("2003-01-01 01:00:00", periods=2, freq="12h")
-            msg = "PeriodIndex.format is deprecated"
-            with tm.assert_produces_warning(FutureWarning, match=msg):
-                formatted = per.format(date_format="%y é")
-            assert formatted[0] == "03 é"
-            assert formatted[1] == "03 é"
-
-    @pytest.mark.parametrize(
-        "locale_str",
-        [
-            pytest.param(None, id=str(locale.getlocale())),
-            "it_IT.utf8",
-            "it_IT",  # Note: encoding will be 'ISO8859-1'
-            "zh_CN.utf8",
-            "zh_CN",  # Note: encoding will be 'gb2312'
-        ],
-    )
-    def test_period_custom_locale_directive(self, locale_str):
-        # GH#46319 locale-specific directive leads to non-utf8 c strftime char* result
-
-        # Skip if locale cannot be set
-        if locale_str is not None and not tm.can_set_locale(locale_str, locale.LC_ALL):
-            pytest.skip(f"Skipping as locale '{locale_str}' cannot be set on host.")
-
-        # Change locale temporarily for this test.
-        with tm.set_locale(locale_str, locale.LC_ALL) if locale_str else nullcontext():
-            # Get locale-specific reference
-            am_local, pm_local = get_local_am_pm()
-
-            # Scalar
-            per = pd.Period("2018-03-11 13:00", freq="h")
-            assert per.strftime("%p") == pm_local
-
-            # Index
-            per = pd.period_range("2003-01-01 01:00:00", periods=2, freq="12h")
-            msg = "PeriodIndex.format is deprecated"
-            with tm.assert_produces_warning(FutureWarning, match=msg):
-                formatted = per.format(date_format="%y %I:%M:%S%p")
-            assert formatted[0] == f"03 01:00:00{am_local}"
-            assert formatted[1] == f"03 01:00:00{pm_local}"
-
-
-class TestDatetimeIndexFormat:
-    def test_datetime(self):
-        msg = "DatetimeIndex.format is deprecated"
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            formatted = pd.to_datetime([datetime(2003, 1, 1, 12), NaT]).format()
-        assert formatted[0] == "2003-01-01 12:00:00"
-        assert formatted[1] == "NaT"
-
-    def test_date(self):
-        msg = "DatetimeIndex.format is deprecated"
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            formatted = pd.to_datetime([datetime(2003, 1, 1), NaT]).format()
-        assert formatted[0] == "2003-01-01"
-        assert formatted[1] == "NaT"
-
-    def test_date_tz(self):
-        dti = pd.to_datetime([datetime(2013, 1, 1)], utc=True)
-        msg = "DatetimeIndex.format is deprecated"
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            formatted = dti.format()
-        assert formatted[0] == "2013-01-01 00:00:00+00:00"
-
-        dti = pd.to_datetime([datetime(2013, 1, 1), NaT], utc=True)
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            formatted = dti.format()
-        assert formatted[0] == "2013-01-01 00:00:00+00:00"
-
-    def test_date_explicit_date_format(self):
-        dti = pd.to_datetime([datetime(2003, 2, 1), NaT])
-        msg = "DatetimeIndex.format is deprecated"
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            formatted = dti.format(date_format="%m-%d-%Y", na_rep="UT")
-        assert formatted[0] == "02-01-2003"
-        assert formatted[1] == "UT"
-
-
-class TestDatetimeIndexUnicode:
-    def test_dates(self):
-        text = str(pd.to_datetime([datetime(2013, 1, 1), datetime(2014, 1, 1)]))
-        assert "['2013-01-01'," in text
-        assert ", '2014-01-01']" in text
-
-    def test_mixed(self):
-        text = str(
-            pd.to_datetime(
-                [datetime(2013, 1, 1), datetime(2014, 1, 1, 12), datetime(2014, 1, 1)]
-            )
-        )
-        assert "'2013-01-01 00:00:00'," in text
-        assert "'2014-01-01 00:00:00']" in text
-
-
-class TestStringRepTimestamp:
-    def test_no_tz(self):
-        dt_date = datetime(2013, 1, 2)
-        assert str(dt_date) == str(Timestamp(dt_date))
-
-        dt_datetime = datetime(2013, 1, 2, 12, 1, 3)
-        assert str(dt_datetime) == str(Timestamp(dt_datetime))
-
-        dt_datetime_us = datetime(2013, 1, 2, 12, 1, 3, 45)
-        assert str(dt_datetime_us) == str(Timestamp(dt_datetime_us))
-
-        ts_nanos_only = Timestamp(200)
-        assert str(ts_nanos_only) == "1970-01-01 00:00:00.000000200"
-
-        ts_nanos_micros = Timestamp(1200)
-        assert str(ts_nanos_micros) == "1970-01-01 00:00:00.000001200"
-
-    def test_tz_pytz(self):
-        dt_date = datetime(2013, 1, 2, tzinfo=pytz.utc)
-        assert str(dt_date) == str(Timestamp(dt_date))
-
-        dt_datetime = datetime(2013, 1, 2, 12, 1, 3, tzinfo=pytz.utc)
-        assert str(dt_datetime) == str(Timestamp(dt_datetime))
-
-        dt_datetime_us = datetime(2013, 1, 2, 12, 1, 3, 45, tzinfo=pytz.utc)
-        assert str(dt_datetime_us) == str(Timestamp(dt_datetime_us))
-
-    def test_tz_dateutil(self):
-        utc = dateutil.tz.tzutc()
-
-        dt_date = datetime(2013, 1, 2, tzinfo=utc)
-        assert str(dt_date) == str(Timestamp(dt_date))
-
-        dt_datetime = datetime(2013, 1, 2, 12, 1, 3, tzinfo=utc)
-        assert str(dt_datetime) == str(Timestamp(dt_datetime))
-
-        dt_datetime_us = datetime(2013, 1, 2, 12, 1, 3, 45, tzinfo=utc)
-        assert str(dt_datetime_us) == str(Timestamp(dt_datetime_us))
-
-    def test_nat_representations(self):
-        for f in (str, repr, methodcaller("isoformat")):
-            assert f(NaT) == "NaT"
 
 
 @pytest.mark.parametrize(
